refactor(web-mobile): move createTimerPool to a leaf module, adopt in admin consoles
Extract createTimerPool from app.jsx into web-mobile/js/timer_pool.jsx
and migrate the two remaining hand-rolled delete-on-fire timer Sets
(admin.jsx competition-view effect, admin_shiaijo.jsx court-feed
effect) onto it, so the self-pruning pattern lives in exactly one
place. The pool must NOT stay in app.jsx: admin modules importing
app.jsx would evaluate it a second time under a different URL (script
tag app.js?v=N vs bare import app.js), the known double-load class.

Both admin sites keep their cancelled/compId guard semantics: the
guard runs inside the scheduled callback exactly as before; only the
Set bookkeeping moved into the pool.

Verified: 2166 vitest + eslint clean, make go/test green, and native-
ESM browser checks on all three surfaces (/display?court=A,
/admin/competition/c1, /admin/shiaijo/A) with live SSE events flowing
through each migrated pool: queue re-sorted live, zero console errors.
